tests/driver_netdev_common: add compile-test for netif drivers
