+++ tests/driver_at86rf215/Makefile
@@ -1,28 +1,7 @@
 BOARD ?= openmote-b
-include ../Makefile.tests_common
-
-# Modules to include:
-USEMODULE += shell
-USEMODULE += shell_commands
-USEMODULE += ps
 
 # the radio driver to test
 USEMODULE += at86rf215
 GNRC_NETIF_NUMOF ?= 2
 
-# gnrc is a meta module including all required, basic gnrc networking modules
-USEMODULE += gnrc
-
-# automatically initialize the network interface
-USEMODULE += auto_init_gnrc_netif
-
-# shell command to send L2 packets with a simple string
-USEMODULE += gnrc_txtsnd
-
-# the application dumps received packets to stdout
-USEMODULE += gnrc_pktdump
-
-include $(RIOTBASE)/Makefile.include
-
-# Set a custom channel if needed
-include $(RIOTMAKE)/default-radio-settings.inc.mk
+include ../driver_netdev_common/Makefile
