+++ tests/driver_at86rf215/main.c
@@ -1,39 +1,1 @@
-/*
- * Copyright (C) 2020 ML!PA Consulting GmbH
- *
- * This file is subject to the terms and conditions of the GNU Lesser
- * General Public License v2.1. See the file LICENSE in the top level
- * directory for more details.
- */
-
-/**
- * @ingroup     tests
- * @{
- *
- * @file
- * @brief       Test application for network device drivers
- *
- * @author      Benjamin Valentin <benjamin.valentin@ml-pa.com>
- * @}
- */
-
-#include "thread.h"
-#include "shell.h"
-#include "shell_commands.h"
-
-#include "net/gnrc/pktdump.h"
-#include "net/gnrc.h"
-
-int main(void)
-{
-    /* enable pktdump output */
-    gnrc_netreg_entry_t dump = GNRC_NETREG_ENTRY_INIT_PID(GNRC_NETREG_DEMUX_CTX_ALL,
-                                                          gnrc_pktdump_pid);
-    gnrc_netreg_register(GNRC_NETTYPE_UNDEF, &dump);
-
-    /* start the shell */
-    char line_buf[SHELL_DEFAULT_BUFSIZE];
-    shell_run(NULL, line_buf, SHELL_DEFAULT_BUFSIZE);
-
-    return 0;
-}
+../driver_netdev_common/main.c
